requirements.txt
torch==2.2.0
torchvision==0.8.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
argon2==0.1.10
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
attrs==23.2.0
barfi==0.7.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
exceptiongroup==1.2.1
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib-resources==6.4.0
jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
networkx==3.1
numpy==1.24.4
openai==1.35.7
packaging==24.1
pandas==2.0.3
pillow==10.3.0
pip==19.3.1
pkgutil-resolve-name==1.3.10
protobuf==5.27.2
pyarrow==16.1.0
pycparser==2.22
pydantic==2.7.4
pydantic-core==2.18.4
pydeck==0.9.1
pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.1
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
setuptools==41.6.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sounddevice==0.4.7
streamlit==1.36.0
streamlit-lottie==0.0.5
streamlit-option-menu==0.3.13
tenacity==8.4.2
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.4
trycourier==6.1.0
typing-extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
watchdog==4.0.1
wavio==0.0.9
websocket-client==1.8.0
zipp==3.19.2

